wards ocd

+++ src/DssDirectDepositHub.sol
@@ -71,7 +71,7 @@ interface DssDirectDepositPoolLike {
 contract DssDirectDepositHub {
 
     // --- Auth ---
-    mapping (address => uint) public wards;
+    mapping (address => uint256) public wards;
     function rely(address usr) external auth {
         wards[usr] = 1;
         emit Rely(usr);
