+++ src/pools/DssDirectDepositAaveDaiPool.sol
@@ -71,7 +71,7 @@ interface DssDirectDepositPlanLike {
 contract DssDirectDepositAaveDaiPool {
 
     // --- Auth ---
-    mapping (address => uint) public wards;
+    mapping (address => uint256) public wards;
     function rely(address usr) external auth {
         wards[usr] = 1;
